add batching to scraper and made database a resource

+++ backend/scrape.py
@@ -13,18 +13,17 @@ from selenium.webdriver.common.by import By
 from selenium import webdriver
 from dotenv import load_dotenv
 import os
-from tqdm import tqdm
 from collections import deque
 from concurrent.futures import ThreadPoolExecutor
-from concurrent.futures import ProcessPoolExecutor
+from numpy import random
 
 INPUT_FILE = 'mtgdataset/data/autotagging.tsv'
 TRACKS, TAGS, EXTRA = read_file(INPUT_FILE)
-DB = Database('first_100', True)
+DB = Database('full_jamendo', True)
 
 
 def download_and_post(download_dir, song_metadata,
-                      download_suffix=".crdownload", wait_time=10, start_buffer=0):
+                      download_suffix=".crdownload", wait_time=10, start_buffer=0.5):
     # wait for download
     filename_prefix = song_metadata['title'].replace(" ", "_")
     filepath = None
@@ -74,16 +73,14 @@ def download_and_post(download_dir, song_metadata,
                        instrument=track['instrument'],
                        moodtheme=track['mood/theme'],
                        description=song_metadata['description'])
-
     DB.post_songs([song_object])
-
-    # clean-up
-    os.remove(current_file)
+    if os.path.exists(current_file):
+        os.remove(current_file)
     # print(f"Downloaded and posted {song_metadata['title']} in {t:.2f} sectionsfrom directory {current_file}!")
 
 
-def download_song(track_id_list, download_wait_time=30, min_threads=0, max_threads=2, process=1):
-    download_dir = os.path.abspath("backend/temp_audio")
+def download_song(track_id_list, download_wait_time=30, min_threads=1, max_threads=2, process=1):
+    download_dir = os.path.abspath(f"backend/temp_audio/{process}")
     if not os.path.exists(download_dir):
         os.makedirs(download_dir)
     caps = DesiredCapabilities.CHROME
@@ -99,6 +96,7 @@ def download_song(track_id_list, download_wait_time=30, min_threads=0, max_threa
     driver = webdriver.Chrome(options=chrome_options)
 
     try:
+        idx = 0
         driver.get("https://www.jamendo.com/start")
         login_button = WebDriverWait(driver, 10).until(
             EC.presence_of_element_located(
@@ -141,12 +139,12 @@ def download_song(track_id_list, download_wait_time=30, min_threads=0, max_threa
         driver.execute_script("arguments[0].click();", submit_button)
 
         # Wait a bit to make sure that you are logged in
-        time.sleep(1)
+        time.sleep(2)
         # print("Login completed!")
         # print("Logged in successfully!")
         with ThreadPoolExecutor(max_workers=max_threads) as executor:
             futures = deque()
-            for id in track_id_list:
+            for idx, id in enumerate(track_id_list):
                 # Define the url to the track_id
                 url = "https://www.jamendo.com/track/" + str(id)
 
@@ -203,17 +201,17 @@ def download_song(track_id_list, download_wait_time=30, min_threads=0, max_threa
                 if len(futures) >= max_threads:
                     while len(futures) > min_threads:
                         futures.popleft().result()
-
-        # Next step: Format download into song objects!
-        # Push the song object to the database!
-        DB.save_db()
-
+        driver.quit()
     except Exception as e:
         print(f"An error occurred: {e}")
         traceback.print_exc()
-
-    finally:
         driver.quit()
+        move_on = random.randint(0, 1)
+        download_song(track_id_list[idx+1:], download_wait_time=download_wait_time,
+                      min_threads=min_threads, max_threads=max_threads, process=process)
+    finally:
+        for file in os.listdir(download_dir):
+            os.remove(os.path.join(download_dir, file))
 
 
 def download_song_parallel(track_id_list, max_processes=10):
@@ -226,16 +224,20 @@ def download_song_parallel(track_id_list, max_processes=10):
                 download_song, track_id_list[start:end], process=i+1))
         for future in futures:
             future.result()
+    DB.save_db()
 
 
 if __name__ == "__main__":
     track_ids = list(TRACKS.keys())
-    track_id_list = track_ids[0:100]
+    track_id_list = track_ids
 
     # Run the download_song feature using this function
     start_time = time.time()
-
-    download_song_parallel(track_id_list, max_processes=10)
+    batches = 100
+    for i in range(15, batches):
+        start = (i*len(track_id_list))//batches
+        end = ((i+1)*len(track_id_list))//batches
+        download_song_parallel(track_id_list[start:end], max_processes=16)
 
     end_time = time.time()
     elapsed_time = end_time - start_time
